requirements.txt
asgiref==3.4.1
bleach==3.3.0
Brotli==1.0.9
cffi==1.14.6
crispy-bootstrap5==0.6
cssselect2==0.4.1
Django==4.0
django-crispy-forms==1.13.0
django-markdownify==0.9.0
djangorestframework==3.13.1
fonttools==4.26.2
html5lib==1.1
Markdown==3.3.4
packaging==21.0
Pillow==8.3.0
pycparser==2.20
pydyf==0.0.3
pyparsing==2.4.7
pyphen==0.11.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
weasyprint==53.0
webencodings==0.5.1
zopfli==0.1.8
